+++ src/main/kotlin/br/com/gamemods/j2nwc/WorldConverterCLI.kt
@@ -17,8 +17,6 @@ object WorldConverterCLI {
      */
     @JvmStatic
     fun main(args: Array<String>) {
-        checkIds()
-
         val jarFileName = JavaChunk::class.java.protectionDomain.codeSource.location.toURI().path.substringAfterLast('/')
 
         val cli = CommandLineInterface("java -jar $jarFileName")
@@ -32,6 +30,14 @@ object WorldConverterCLI {
         val keepCustomHeads by cli.flagArgument(
             "--keep-custom-heads",
             "Makes the tool convert player heads with custom skins as regular player heads. The default behaviour is to remove them."
+        )
+        val target by cli.flagValueArgument(
+            "--target",
+            "nukkit-implementation",
+            "The target output format. Valid values are nukkit (for Nukkit v1.X), nukkit2 (for Nukkit v2.X), " +
+                    "powernukkit (for PowerNukkit v1.X) and powernukkit2 (for PowerNukkit v2.X)",
+            "nukkit",
+            { it.trim().toLowerCase() }
         )
 
         try {
@@ -52,6 +58,23 @@ object WorldConverterCLI {
             exitProcess(3)
         }
 
+        val targetType = when (target) {
+            "nukkit" -> WorldConverter.TargetType.NUKKIT
+            "powernukkit" -> WorldConverter.TargetType.POWER_NUKKIT
+            "nukkit2", "powernukkit2" -> {
+                System.err.println("Direct world conversion to LevelDB isn't supported yet, sorry. " +
+                        "Try ${target.substring(0, target.length - 1)}, the server should update to the 2.0 automatically.")
+                exitProcess(7)
+            }
+            else -> {
+                System.err.println("Unknown target type $target, valid targets are nukkit, nukkit2, powernukkit, powernukkit2")
+                cli.printHelp()
+                exitProcess(8)
+            }
+        }
+
+        checkIds(targetType)
+
         val regionLimit = if (regionsArg == null) {
             emptyList()
         } else {
@@ -68,18 +91,19 @@ object WorldConverterCLI {
             }
         }
 
-        val fromPath = Paths.get(from)
+        val fromPath = Paths.get(from!!)
         if (!Files.isDirectory(fromPath)) {
             System.err.println("$from is not a folder!")
             cli.printHelp()
             exitProcess(4)
         }
 
-        val toPath = Paths.get(to)
+        val toPath = Paths.get(to!!)
         try {
             WorldConverter(fromPath.toFile(), toPath.toFile()).apply {
                 regions = regionLimit.toMutableSet()
                 skipSkinHeads = !keepCustomHeads
+                this.targetType = targetType
                 convert()
             }
             println("The world has been converted successfully")